# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Magic Garden API
  version: 2.0.0
  description: |
    **Unofficial API** for Magic Garden game data that dynamically extracts and provides game content.

    ## Overview

    This API provides two types of data:
    - **Static data**: Extracted from the game's minified bundle (plants, pets, items, etc.)
    - **Live data**: Real-time data from the game's WebSocket (shops, weather)

    ## Quick Start

    ### Get All Game Data
    ```bash
    curl https://mg-api.ariedam.fr/data
    ```

    ### Get Specific Plant
    ```bash
    curl https://mg-api.ariedam.fr/data/plants | jq '.Carrot'
    ```

    ### Stream Live Updates
    ```javascript
    const eventSource = new EventSource('https://mg-api.ariedam.fr/live/stream');
    eventSource.addEventListener('weather', (event) => {
      const data = JSON.parse(event.data);
      console.log('Weather:', data.weather);
    });
    eventSource.addEventListener('shops', (event) => {
      const shops = JSON.parse(event.data);
      console.log('Seed shop:', shops.seed);
    });
    ```

    ## Key Features

    - **Auto-updating**: Data automatically refreshes from the game bundle
    - **Real-time**: WebSocket connection for live shop and weather data
    - **Sprites included**: All data includes sprite URLs ready to use
    - **Server-Sent Events**: Real-time streaming via SSE
    - **Smart caching**: Optimized performance with automatic cache invalidation
...

    ## Data Structure

...

    ## Rate Limiting
...
servers:
...
paths:
...
  /assets/sprite-data:
    get:
      tags:
        - Assets
      summary: Get sprites data
      description: |
        Returns sprite atlas metadata with powerful filtering options.

        **Query parameters:**
        - `search`: Filter sprites by name (case-insensitive partial match)
        - `cat`: Filter by category (seeds, plants, pets, etc.)
        - `full=1`: Include detailed sprite information (anchors, frames, source sizes)
        - `flat=1`: Return flat array instead of grouped by category

        **Use cases:**
        - Find sprite URLs for specific items
        - Browse available sprites by category
        - Get sprite dimensions and atlas coordinates
        - Build sprite catalogs

        **Response notes:**
        - `baseUrl` points to the game's versioned asset base, not the API host.
        - `url` points to the atlas PNG; use `/assets/sprites` for individual sprite PNGs.

        **Examples:**
        - `/assets/sprite-data?search=Carrot` - Find all Carrot sprites
        - `/assets/sprite-data?cat=seeds` - Get all seed sprites
        - `/assets/sprite-data?search=Bunny&full=1` - Get detailed Bunny sprite info
        - `/assets/sprite-data?cat=pets&flat=1` - Flat list of pet sprites
      parameters:
        - name: full
          in: query
          description: Include full sprite information (anchors, source sizes)
          schema:
            type: string
            enum: ['1']
          example: '1'
        - name: search
          in: query
          description: Filter sprites by ID (partial match, case-insensitive)
          schema:
            type: string
          example: 'Carrot'
        - name: cat
          in: query
          description: Filter by category
          schema:
            type: string
            enum: ["seeds", "plants", "tallPlants", "mutations", "pets", "decor", "items", "objects", "ui", "animations", "weather", "tiles", "winter"]
          example: 'seeds'
        - name: flat
          in: query
          description: Return flat list instead of grouped by category
          schema:
            type: string
            enum: ['1']
          example: '1'
      responses:
        '200':
          description: Sprites data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpritesResponse'
              examples:
                grouped:
                  summary: Grouped response (default)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 3
                    categories:
                      - cat: "seeds"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                          - id: "Cabbage"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - cat: "plants"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                flat:
                  summary: Flat response (with flat=1)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 3
                    items:
                      - id: "Carrot"
                        cat: "seeds"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - id: "Cabbage"
                        cat: "seeds"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - id: "Carrot"
                        cat: "plants"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                fullDetails:
                  summary: Full details (with full=1)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 1
                    categories:
                      - cat: "seeds"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                            frame:
                              x: 0
                              y: 0
                              w: 64
                              h: 64
                            anchor:
                              x: 0.5
                              y: 0.5
                            sourceSize:
                              w: 64
                              h: 64
  /assets/cosmetics:
    get:
      tags:
        - Assets
      summary: Get cosmetics data
      description: |
        Returns all cosmetic items for player avatar customization.

        **Categories include:**
        - Hats
        - Outfits
        - Accessories
        - Hair styles
        - Face customization

        **Query parameters:**
        - `full=1`: Include detailed sprite information (frames, anchors)

        **Use cases:**
        - Build character customization UI
        - Display cosmetic shop
        - Show available customization options

        **Response notes:**
        - `baseUrl` points to the game's versioned asset base, not the API host.
      parameters:
        - name: full
          in: query
          description: Include full cosmetic information (sprite frames, anchors)
          schema:
            type: string
            enum: ['1']
          example: '1'
      responses:
        '200':
          description: Cosmetics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CosmeticsResponse'
              examples:
                cosmeticsData:
                  summary: Example cosmetics
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 25
                    categories:
                      - cat: "hats"
                        items:
                          - id: "StrawHat"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/hats/StrawHat.png"
                          - id: "WizardHat"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/hats/WizardHat.png"
                      - cat: "outfits"
                        items:
                          - id: "FarmerOutfit"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/outfits/FarmerOutfit.png"
...
  /assets/sprites:
    get:
      tags:
        - Assets
      summary: Full sprite catalog with ready-to-use PNG URLs
      description: |
        Returns the complete catalog of exported sprites, grouped by category,
        with a ready-to-use PNG URL for each one. Around 600 sprites in total.

        Want a subset? Filter with `cat` (one category) and/or `search`
        (case-insensitive match on the sprite name).

        **Available categories:**
        - `seeds` - Seed sprites
        - `plants` - Plant growth stages
        - `tallPlants` - Tall plant varieties
        - `mutations` - Mutation variants
        - `pets` - Pet companions
        - `decor` - Garden decorations
        - `items` - Tools and consumables
        - `objects` - Interactive objects
        - `ui` - UI elements
        - `animations` - Animation frames
        - `weather` - Weather effects
        - `tiles` - Ground tiles
        - `winter` - Winter seasonal sprites

        **Use cases:**
        - Discover every sprite the game currently ships
        - Build sprite pickers or galleries without hardcoding names
        - Find a sprite's exact filename before calling `/assets/sprites/{category}/{name}`

        **Examples:**
        - `/assets/sprites` - everything
        - `/assets/sprites?cat=pets` - all pet sprites
        - `/assets/sprites?search=carrot` - anything named like "carrot", across categories

        **Response notes:**
        - `baseUrl` reflects the public API base used for sprite URLs (`SPRITES_BASE_URL`).
        - URLs include a `?v=<gameVersion>` cache-busting parameter.
        - `categories` always lists every valid category, even when you filter.
      parameters:
        - name: cat
          in: query
          required: false
          schema:
            type: string
            enum: [seeds, plants, tallPlants, mutations, pets, decor, items, objects, ui, animations, weather, tiles, winter]
          description: Only return sprites from this category
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on sprite names
          example: "carrot"
      responses:
        '200':
          description: Sprite catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Number of sprites in the response (after filters)
                    example: 600
                  baseUrl:
                    type: string
                    description: Base URL for the API
                    example: "https://mg-api.ariedam.fr"
                  categories:
                    type: array
                    items:
                      type: string
                    description: Every valid category (not affected by filters)
                    example: ["animations", "decor", "items", "mutations", "objects", "pets", "plants", "seeds", "tallPlants", "tiles", "ui", "weather", "winter"]
                  sprites:
                    type: object
                    description: Sprites grouped by category, each with a ready-to-use PNG URL
                    additionalProperties:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                            example: "Carrot"
                          url:
                            type: string
                            example: "https://mg-api.ariedam.fr/assets/sprites/seeds/Carrot.png?v=642"
              examples:
                filtered:
                  summary: Filtered catalog (?search=carrot)
                  value:
                    count: 3
                    baseUrl: "https://mg-api.ariedam.fr"
                    categories: ["animations", "decor", "items", "mutations", "objects", "pets", "plants", "seeds", "tallPlants", "tiles", "ui", "weather", "winter"]
                    sprites:
                      plants:
                        - name: "BabyCarrot"
                          url: "https://mg-api.ariedam.fr/assets/sprites/plants/BabyCarrot.png?v=642"
                        - name: "Carrot"
                          url: "https://mg-api.ariedam.fr/assets/sprites/plants/Carrot.png?v=642"
                      seeds:
                        - name: "Carrot"
                          url: "https://mg-api.ariedam.fr/assets/sprites/seeds/Carrot.png?v=642"
        '400':
          description: Unknown category in `cat`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /assets/animations:
    get:
      tags:
        - Assets
      summary: Catalog of animated pet loops (WebP/GIF)
      description: |
        Pets are the only creatures the game renders as vectors rather than
        sprites, so beyond the still PNG served by `/assets/sprites/pets/*`
        this API can serve them **animated**: one looping WebP (or GIF) per
        species and per state.

        Loops are pre-rendered whenever the game ships a new pet file - never
        on request - and served as plain files. Drop the URL in an `<img>` tag
        and it plays; there is no runtime to embed.

        **Clips:**
        - `idle` - the resting loop, 7 s
        - `walk` - one step cycle, ~0.6 s
        - `eat` - eating, 1 s
        - `sleep` - sleeping, 8 s

        Not every species exposes every clip, and the two weather variants
        (`FireHorseActive`, `ThunderWolfActive`) only ship `idle`.

        **Filters:** `cat`, `name` (exact species), `clip`, `search`
        (case-insensitive substring on the species name).

        **Response notes:**
        - Every loop is rendered to the same subject height, so species stay
          visually comparable - a snail and a horse come out the same size.
        - `url` is the link, in the primary format (WebP by default); `format`
          names it, and any additional format appears under its own key
          (`gif`).
        - `anchor` places the sprite like the game does: `x` is always 0.5,
          `y` marks the ground line under the pet's feet.
        - `source.generatedAt` tells you when the loops were last rendered, and
          `source.riveUrl` which game file they came from.
        - Not to be confused with the `animations` **sprite** category
          (`/assets/sprites/animations/*.png`), which holds still frames from
          the game's atlases.
      parameters:
        - name: cat
          in: query
          required: false
          schema:
            type: string
            enum: [pets]
          description: Only return animations from this category
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Exact species name (case-insensitive)
          example: "Chicken"
        - name: clip
          in: query
          required: false
          schema:
            type: string
            enum: [idle, walk, eat, sleep]
          description: Only return this clip
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on species names
          example: "horse"
      responses:
        '200':
          description: Animation catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 118
                  baseUrl:
                    type: string
                    example: "https://mg-api.ariedam.fr"
                  categories:
                    type: array
                    items:
                      type: string
                    example: ["pets"]
                  formats:
                    type: array
                    items:
                      type: string
                    description: Formats currently generated
                    example: ["webp"]
                  source:
                    type: object
                    properties:
                      riveUrl:
                        type: string
                        description: Game file the loops were rendered from
                      generatedAt:
                        type: string
                        format: date-time
                  animations:
                    type: object
                    description: Animations grouped by category
                    additionalProperties:
                      type: array
                      items:
                        $ref: '#/components/schemas/PetAnimation'
              examples:
                filtered:
                  summary: One species (?name=Chicken)
                  value:
                    count: 4
                    baseUrl: "https://mg-api.ariedam.fr"
                    categories: ["pets"]
                    formats: ["webp"]
                    source:
                      riveUrl: "https://magicgarden.gg/runtime-assets/pets.bfec8b856f6e2b0e4a83.riv"
                      generatedAt: "2026-08-04T14:31:07.412Z"
                    animations:
                      pets:
                        - category: "pets"
                          name: "Chicken"
                          clip: "idle"
                          timeline: "Pet_Idle"
                          url: "https://mg-api.ariedam.fr/assets/animations/pets/Chicken_idle.webp?v=824"
                          format: "webp"
                          width: 212
                          height: 257
                          frames: 210
                          fps: 30
                          durationMs: 7000
                          anchor:
                            x: 0.5
                            y: 1
        '400':
          description: Unknown category in `cat`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /assets/rive:
    get:
      tags:
        - Assets
      summary: The game's Rive (vector) files and what they contain
      description: |
        Everything that moves in the game is vector art, rendered live by a Rive
        runtime: the pets, but also the animated decor, the thought bubble, the
        coin and the gift box. This endpoint publishes the files themselves and
        an inventory of what is inside them.

        **Why you may want the file rather than the images.** `pets.riv` is
        3 MB and holds all 28 species with about 660 timelines, at any
        resolution and frame rate. The equivalent coverage as pre-encoded WebP
        would run to hundreds of megabytes. If your client can run the Rive
        runtime, play the source; if it can only show a picture, use
        `/assets/animations`.

        **What the inventory is for.** Artboard names, state machine names,
        timeline names and - above all - **the type of every state machine
        input**. Half of a pet's inputs are *triggers* (`walk`, `eat`,
        `petted`), not booleans (`sleep`, `fire`, `thunder`), and asking for the
        wrong one fails silently. All of it changes with game updates, so
        nothing here should ever be hardcoded by a client.

        **Notes:**
        - `url` goes through `/assets/proxy`, which adds the CORS headers
          magicgarden.gg does not send - a browser cannot fetch `origin`
          directly.
        - `origin` is versioned by content hash, so it changes on every update.
        - `loadable: false` means our pinned runtime cannot read the file.
          `avatar.riv` is permanently in this state: it references assets stored
          outside the file (the cosmetics bundle) and never finishes loading.
        - The inventory is refreshed during sprite sync, never per request:
          inspecting a file costs a download and a WASM pass.
      parameters:
        - name: key
          in: query
          required: false
          schema:
            type: string
          description: Return a single file, with full artboard detail
          example: "pets"
        - name: full
          in: query
          required: false
          schema:
            type: string
            enum: ["1"]
          description: Include every artboard's timelines and state machines
      responses:
        '200':
          description: Rive inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 6
                  baseUrl:
                    type: string
                    example: "https://mg-api.ariedam.fr"
                  generatedAt:
                    type: string
                    format: date-time
                    description: When the inventory was last refreshed
                  files:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/RiveFile'
              examples:
                summary:
                  summary: Default view
                  value:
                    count: 6
                    baseUrl: "https://mg-api.ariedam.fr"
                    generatedAt: "2026-08-04T18:25:17.004Z"
                    files:
                      pets:
                        key: "pets"
                        aliases: ["rive/pets.riv", "rive/pets"]
                        bundle: "default"
                        url: "https://mg-api.ariedam.fr/assets/proxy?url=https%3A%2F%2Fmagicgarden.gg%2Fruntime-assets%2Fpets.bfec8b856f6e2b0e4a83.riv"
                        origin: "https://magicgarden.gg/runtime-assets/pets.bfec8b856f6e2b0e4a83.riv"
                        bytes: 3085491
                        loadable: true
                        artboardCount: 29
                        timelineCount: 638
                        artboards: ["Pets", "Pig", "Horse", "Chicken"]
                      avatar:
                        key: "avatar"
                        bundle: "default"
                        loadable: false
                        error: "Rive load timed out after 20000ms"
        '404':
          description: Unknown `key`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /assets/sprites/composed:
    get:
      tags:
        - Assets
      summary: Get pre-composed sprite with mutations
      description: |
        Returns a single PNG with all mutation layers pre-rendered on top of the base sprite.

        The pipeline applies (bottom → top):
        1. Color filters (tint/gradient per mutation)
        2. Base sprite
        3. Mutation icons
        4. Tall-plant overlays (masked to base silhouette, tall plants only)
        5. Floating icons (Dawnlit, Ambershine, Dawncharged, Ambercharged)

        The output canvas expands to fit all layers, so icons that extend past the base sprite
        bounds are never clipped.

        **Mutations supported:** `Gold`, `Rainbow`, `Wet`, `Chilled`, `Frozen`, `Thunderstruck`,
        `Dawnlit`, `Ambershine`, `Dawncharged`, `Ambercharged`

        Unknown mutation ids are silently ignored.

        **Examples:**
        - `/assets/sprites/composed?key=sprite/plant/Cactus&mutations=Rainbow,Wet`
        - `/assets/sprites/composed?key=sprite/plant/Carrot&mutations=Gold`
        - `/assets/sprites/composed?key=sprite/pet/Bunny&mutations=Dawnlit,Ambershine`

        **Caching:** `public, max-age=86400, stale-while-revalidate=3600`
      parameters:
        - name: key
          in: query
          required: true
          description: Full atlas key of the base sprite (e.g. `sprite/plant/Cactus`, `sprite/plant/Carrot`, `sprite/pet/Bunny`)
          schema:
            type: string
          example: "sprite/plant/Cactus"
        - name: mutations
          in: query
          required: false
          description: Comma-separated list of mutation ids (order-insensitive, deduplicated server-side)
          schema:
            type: string
          example: "Rainbow,Wet"
      responses:
        '200':
          description: Composed PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          description: Missing or invalid `key` parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingKey:
                  summary: Missing key
                  value:
                    error:
                      code: "BAD_REQUEST"
                      message: "Missing required query param: key"
                invalidKey:
                  summary: Invalid key format
                  value:
                    error:
                      code: "BAD_REQUEST"
                      message: "Invalid key format"
        '404':
          description: Base sprite not found in atlas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  summary: Sprite not found
                  value:
                    error:
                      code: "NOT_FOUND"
                      message: "Sprite not found: sprite/plant/Unknown"
...
components:
  schemas:
...
    RiveFile:
      type: object
      description: One Rive file from the game's manifest, and what it contains
      properties:
        key:
          type: string
          description: Stable key, taken from the filename (not the alias, which the game can rename)
          example: "pets"
        aliases:
          type: array
          items:
            type: string
          example: ["rive/pets.riv", "rive/pets"]
        bundle:
          type: string
          description: Manifest bundle declaring it - only `pets` and `avatar` live in `default`
          example: "default"
        url:
          type: string
          format: uri
          description: Fetchable from a browser, through `/assets/proxy`
        origin:
          type: string
          format: uri
          description: Upstream URL, versioned by content hash
        bytes:
          type: integer
          example: 3085491
        loadable:
          type: boolean
          description: Whether the pinned runtime can read the file
          example: true
        error:
          type: string
          description: Present when `loadable` is false
        defaultArtboard:
          type: string
          description: Container artboard the game never renders directly
          example: "Pets"
        artboardCount:
          type: integer
          example: 29
        timelineCount:
          type: integer
          example: 638
        artboards:
          description: |
            Artboard names by default; full objects with `?full=1` or `?key=`.
          oneOf:
            - type: array
              items:
                type: string
            - type: array
              items:
                $ref: '#/components/schemas/RiveArtboard'
        inspectedAt:
          type: string
          format: date-time
...
    PetAnimation:
      type: object
      description: Catalog entry for one looping animation
      properties:
        category:
          type: string
          example: "pets"
        name:
          type: string
          description: Artboard name (species, or weather variant)
          example: "Chicken"
        clip:
          type: string
          enum: [idle, walk, eat, sleep]
          example: "idle"
        timeline:
          type: string
          description: Rive timeline the loop was rendered from
          example: "Pet_Idle"
        url:
          type: string
          format: uri
          example: "https://mg-api.ariedam.fr/assets/animations/pets/Chicken_idle.webp?v=824"
        format:
          type: string
          description: Format `url` points to
          example: "webp"
        gif:
          type: string
          format: uri
          description: Only present when the server also generates GIFs
        width:
          type: integer
          example: 212
        height:
          type: integer
          example: 257
        frames:
          type: integer
          example: 210
        fps:
          type: number
          example: 30
        durationMs:
          type: integer
          example: 7000
        anchor:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        bytes:
          type: object
          description: File size per format
          additionalProperties:
            type: object
            properties:
              bytes:
                type: integer
...
    SpritesResponse:
      type: object
      properties:
        baseUrl:
          type: string
          description: Game assets base URL (versioned)
        count:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              cat:
                type: string
              items:
                type: array
                items:
                  type: object
    CosmeticsResponse:
      type: object
      properties:
        baseUrl:
          type: string
          description: Game assets base URL (versioned)
        count:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              cat:
                type: string
              items:
                type: array
                items:
                  type: object
...
    Error:
      type: object
      description: Standard error response format
      properties:
        error:
          type: object
          description: Error details
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "SPRITE_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "The requested sprite was not found"
            details:
              type: object
              description: Additional error context (varies by error type)
              additionalProperties: true
              example:
                category: "seeds"
                filename: "NonExistent.png"
      example:
        error:
          code: "SPRITE_NOT_FOUND"
          message: "The requested sprite was not found"
          details:
            category: "seeds"
            filename: "NonExistent.png"
...